requirements.txt
# Core FastAPI dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic[email]==2.5.0
aiohttp==3.9.1

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6

# Database
sqlalchemy==2.0.23

# Gradio for HF Spaces interface
gradio==4.8.0

# HTTP requests
requests==2.31.0
httpx==0.25.2

# File processing
python-docx==1.1.0
PyPDF2==3.0.1
pdf2image==1.16.3
Pillow==10.1.0

# AI/ML dependencies - FIXED VERSIONS FOR COMPATIBILITY
numpy==1.24.4
transformers==4.36.0
torch==2.1.1
torchaudio==2.1.1
torchvision==0.16.1
tokenizers==0.15.0
sentence-transformers==2.2.2

# OCR specific dependencies
PyMuPDF==1.23.5
opencv-python-headless==4.8.1.78

# Text processing
spacy==3.7.2
nltk==3.8.1

# Web scraping - اضافه شده
beautifulsoup4==4.12.2
lxml==4.9.3
selenium==4.15.2

# Utilities
python-dotenv==1.0.0
pathlib2==2.3.7
typing-extensions==4.8.0

# Logging and monitoring
structlog==23.2.0

# Date and time
python-dateutil==2.8.2

# Environment and configuration
pydantic-settings==2.1.0

# Optional: Redis for caching
redis==5.0.1

# Development dependencies
pytest==7.4.3
pytest-asyncio==0.21.1

# Hugging Face specific
huggingface-hub==0.19.4

# Additional utilities
aiofiles==23.2.1
jinja2==3.1.2

# Image processing
scikit-image==0.21.0

# Additional missing dependencies
websockets==11.0.3
python-magic==0.4.27
chardet==5.2.0
protobuf 